+++ Platforms/macOS/App/ZYCandidatePanel.h
@@ -15,6 +15,7 @@
 @property(nonatomic,weak) id<ZYCandidatePanelDelegate> candidateDelegate;
 - (void)updateCandidates:(const ZYCandidate *)items deletable:(const BOOL *)deletable count:(NSUInteger)count selected:(NSUInteger)selected chinese:(BOOL)chinese simplified:(BOOL)simplified;
 - (void)updateWords:(NSArray<NSString*> *)words count:(NSUInteger)count selected:(NSUInteger)selected chinese:(BOOL)chinese simplified:(BOOL)simplified modeLabel:(NSString *)modeLabel;
+- (void)setPreeditText:(NSString *)text;
 - (void)setDeleteMode:(BOOL)enabled;
 - (void)showNearRect:(NSRect)rect;
 - (void)toggleQuickHelp;